stake-contract: refactored tests
