+++ contracts/stake/src/state.rs
@@ -74,7 +74,7 @@ impl StakeState {
 
         rusk_abi::emit(
             "stake",
-            StakeEvent {
+            StakingEvent {
                 public_key: stake.public_key,
                 value: stake.value,
             },
