+++ contracts/stake/tests/common/utils.rs
@@ -1,0 +1,120 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+// Copyright (c) DUSK NETWORK. All rights reserved.
+
+use std::sync::mpsc;
+
+use dusk_pki::ViewKey;
+use dusk_plonk::prelude::*;
+use phoenix_core::transaction::{TreeLeaf, TRANSFER_TREE_DEPTH};
+use phoenix_core::{Note, Transaction};
+use poseidon_merkle::Opening as PoseidonOpening;
+use rusk_abi::Error;
+use rusk_abi::TRANSFER_CONTRACT;
+use rusk_abi::{CallReceipt, ContractError, Session};
+
+const POINT_LIMIT: u64 = 0x100000000;
+
+const H: usize = TRANSFER_TREE_DEPTH;
+const A: usize = 4;
+
+pub fn leaves_from_height(
+    session: &mut Session,
+    height: u64,
+) -> Result<Vec<TreeLeaf>, Error> {
+    let (feeder, receiver) = mpsc::channel();
+
+    session.feeder_call::<_, ()>(
+        TRANSFER_CONTRACT,
+        "leaves_from_height",
+        &height,
+        feeder,
+    )?;
+
+    Ok(receiver
+        .iter()
+        .map(|bytes| rkyv::from_bytes(&bytes).expect("Should return leaves"))
+        .collect())
+}
+
+pub fn update_root(session: &mut Session) -> Result<(), Error> {
+    session
+        .call(TRANSFER_CONTRACT, "update_root", &(), POINT_LIMIT)
+        .map(|r| r.data)
+}
+
+pub fn root(session: &mut Session) -> Result<BlsScalar, Error> {
+    session
+        .call(TRANSFER_CONTRACT, "root", &(), POINT_LIMIT)
+        .map(|r| r.data)
+}
+
+pub fn opening(
+    session: &mut Session,
+    pos: u64,
+) -> Result<Option<PoseidonOpening<(), H, A>>, Error> {
+    session
+        .call(TRANSFER_CONTRACT, "opening", &pos, POINT_LIMIT)
+        .map(|r| r.data)
+}
+
+pub fn prover_verifier(circuit_name: &str) -> (Prover, Verifier) {
+    let circuit_profile = rusk_profile::Circuit::from_name(circuit_name)
+        .expect(&format!(
+            "There should be circuit data stored for {}",
+            circuit_name
+        ));
+    let (pk, vd) = circuit_profile
+        .get_keys()
+        .expect(&format!("there should be keys stored for {}", circuit_name));
+
+    let prover = Prover::try_from_bytes(pk).unwrap();
+    let verifier = Verifier::try_from_bytes(vd).unwrap();
+
+    (prover, verifier)
+}
+
+pub fn filter_notes_owned_by<I: IntoIterator<Item = Note>>(
+    vk: ViewKey,
+    iter: I,
+) -> Vec<Note> {
+    iter.into_iter().filter(|note| vk.owns(note)).collect()
+}
+
+/// Executes a transaction, returning the call receipt
+pub fn execute(
+    session: &mut Session,
+    tx: Transaction,
+) -> Result<CallReceipt<Result<Vec<u8>, ContractError>>, Error> {
+    // Spend the inputs and execute the call. If this errors the transaction is
+    // unspendable.
+    let mut receipt = session.call::<_, Result<Vec<u8>, ContractError>>(
+        TRANSFER_CONTRACT,
+        "spend_and_execute",
+        &tx,
+        tx.fee.gas_limit,
+    )?;
+
+    // Ensure all gas is consumed if there's an error in the contract call
+    if receipt.data.is_err() {
+        receipt.gas_spent = receipt.gas_limit;
+    }
+
+    // Refund the appropriate amount to the transaction. This call is guaranteed
+    // to never error. If it does, then a programming error has occurred. As
+    // such, the call to `Result::expect` is warranted.
+    let refund_receipt = session
+        .call::<_, ()>(
+            TRANSFER_CONTRACT,
+            "refund",
+            &(tx.fee, receipt.gas_spent),
+            u64::MAX,
+        )
+        .expect("Refunding must succeed");
+
+    receipt.events.extend(refund_receipt.events);
+
+    Ok(receipt)
+}
